On case-insensitive file systems, do intelligent deps file handling.
Prototyping for issue #402

Not clear what to do here exactly. One possibility is to not
error on this if a case-insensitive compare is equal. The downsides are:
1.) It doesn't match what happens on case-sensitive file systems
    (they rebuild)
2.) The depfile always stays around with the wrong case unless its
    corresponding source file is touched (renaming a file doesn't
    do that), and the addition depfile strcmp is slower.

Another possibility is to treat the depfile as non-existent when
the case doesn't match. That's closer to what happens on case-
sensitive file systems.

Other open issues:
* The OS-specific code for Windows (and possibly other POSIXen)
* strcasecmp() matches using the current locale, which might not
  match the file system locale.