Improve output when running a regular Windows ninja binary under cygwin or similar.
Previously, ninja would assume that if it's running on Windows, it's running
in cmd.exe and would disable line overprinting if GetConsoleScreenBufferInfo()
failed.

Instead, try to detect if (non-cygwin-built) ninja is running in cygwin.  The
following heuristics are used:

1. TERM needs to be set, and it must contain a value other than "dumb".  This
   isn't set when running in cmd.exe, so ninja in cmd.exe writing to a pipe or
   running in a non-interactive session should still work.
2. COLUMNS needs to be set.  This is used to detect non-interactive cygwins, and
   it's also used as the target width for eliding.

This should not change behavior on POSIX, and it shouldn't change behavior in
cmd.exe (though ninja can now be tricked by setting TERM, COLUMNS, and making
it write to a pipe so that GetConsoleScreenBufferInfo() fails).  It also
shouldn't change behavior on bots that run builds in cygwin since COLUMNS
shouldn't be set there.

In cygwin (and similar: msys, git bash, ...) ninja should now magically do
statusline overprinting, and it also should no longer strip ANSI color codes
from invoked subcommands.

It's not completely perfect: `ninja | cat` in an interactive session will
(incorrectly) print control characters (for line overwriting) to the pipe.
I couldn't find a way around this.  You can manually set TERM=dumb while a
command like this runs if it bothers you.  Having overpriting work in cygwin
seems worth this bug.

This is similar to #659 but without the cygwin-specific code in patch set 1,
and it doesn't require the manual work needed in patch set 2.  Also, that patch
didn't enable eliding.