Fix crash on cyclic rule variable references.
Fixes #902.

This approach does static cycle detection during rule parsing.  This is
nice because there usually aren't many rules, so it's ok if this takes
some time.  Also, it makes it possible to use lexer_.Error() for a nice
error message.

On the flipside, it doesn't work with the current rspfile logic: ninja
currently also checks statically that both rspfile and rspfile_contents
are set on a rule, so one has to write

    rule foo
      rspfile = $rspfile
      rspfile_content = actual content
    build out: foo in
      rspfile = actual.rsp

to work around this static checking  -- and that's a static cycle for
`rspfile`.  We could make the rspfile / rspfile_content check dynamic,
but many ninja files out there probably contain these cycles, so this
isn't a fix that'll work in practice :-(

(On the plus side: yay, this was covered by tests.)